audioadapter: place (Enum)Format in (Enum)PortConfig
Enumerating the Format on a node does not make sense, this is a port or
stream only property.

Move the enum of the formats to the portconfig param, where it is used
to configure the port group.

This also makes it possible later to have multiple port groups on a node
and have the session manager configure and link them individually.