v4l2: handle min/max for continuous frame interval correctly
v4l2 frame intervals instead of frame rates, which is basically the
inverse. For the most part, this is already handled correctly and
numerator and denominator are swapped accordingly.

However, the minimim frame interval is the maximum frame rate so those
need to be swapped as well.

Without this, the minimum frame rate is larger than the maximum frame
rate for v4l2 devices that define a continuous frame interval.